+++ src/main/java/de/samply/annotations/Bridgehead.java
@@ -1,0 +1,4 @@
+package de.samply.annotations;
+
+public @interface Bridgehead {
+}