+++ src/main/java/de/samply/db/repository/ProjectBridgeheadRepository.java
@@ -1,10 +1,15 @@
 package de.samply.db.repository;
 
+import de.samply.db.model.Project;
 import de.samply.db.model.ProjectBridgehead;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface ProjectBridgeheadRepository extends JpaRepository<ProjectBridgehead, Long> {
+
+    Optional<ProjectBridgehead> findFirstByBridgeheadAndProject(String bridgehead, Project project);
 
 }