+++ src/main/java/de/samply/security/OidcProjectUserService.java
@@ -1,7 +1,7 @@
 package de.samply.security;
 
 import de.samply.app.ProjectManagerConst;
-import de.samply.user.OrganisationRole;
+import de.samply.user.roles.OrganisationRole;
 import de.samply.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -63,10 +63,10 @@ public class OidcProjectUserService extends OidcUserService {
     }
 
     private void importNewUsers() {
-        if (sessionUser.getOrganisationRolesNotDependentOnBridgeheads().contains(OrganisationRole.PROJECT_MANAGER_ADMIN)) {
+        if (sessionUser.getUserOrganisationRoles().getRolesNotDependentOnBridgeheads().contains(OrganisationRole.PROJECT_MANAGER_ADMIN)) {
             userService.createProjectManagerAdminUserIfNotExists(sessionUser.getEmail());
         }
-        sessionUser.getBridgeheads().forEach(bridgehead -> sessionUser.getBridgeheadRoles(bridgehead).ifPresent(organisationRoles -> {
+        sessionUser.getBridgeheads().forEach(bridgehead -> sessionUser.getUserOrganisationRoles().getBridgeheadRoles(bridgehead).ifPresent(organisationRoles -> {
             if (organisationRoles.contains(OrganisationRole.BRIDGEHEAD_ADMIN)) {
                 userService.createBridgeheadAdminUserIfNotExists(sessionUser.getEmail(), bridgehead);
             }