+++ src/main/java/de/samply/user/roles/UserOrganisationRoles.java
@@ -1,0 +1,5 @@
+package de.samply.user.roles;
+
+public class UserOrganisationRoles extends UserRoles<OrganisationRole> {
+
+}