+++ src/main/java/de/samply/user/roles/UserProjectRoles.java
@@ -1,0 +1,4 @@
+package de.samply.user.roles;
+
+public class UserProjectRoles extends UserRoles<ProjectRole> {
+}